# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LinkedIn Bot API
  description: |
    API per il controllo e monitoraggio del LinkedIn Bot.
    Autenticazione: API Key (header `x-api-key`) o Basic Auth, oppure session cookie per la dashboard.
  version: 2.1.0
  contact:
    name: LinkedIn Bot
servers:
  - url: http://localhost:3000
    description: Sviluppo locale
security:
  - apiKey: []
  - basicAuth: []
...
paths:
...
  /api/v1/automation/incidents:
    get:
      tags: [Automation API v1]
      summary: Lista incidenti (v1 envelope)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        '200':
          description: Incidents v1
...
  /api/v1/automation/controls/resume:
    post:
      tags: [Automation API v1]
      summary: Resume automazione (v1)
      description: >
        Rilascia la pausa. Questo namespace e' consumato da automazioni (workflow n8n), che non
        hanno l'incidente sotto gli occhi: vale la regola del canale cieco, quindi rilascia SOLO
        una pausa chiesta dall'utente e `force` NON e' disponibile. Per forzare serve la
        dashboard o la CLI.
      responses:
        '200':
          description: Resumed
        '409':
          description: >
            Rifiutata da una protezione attiva (`error.code = CONTROL_ACTION_REJECTED`,
            `error.details.blockedBy` = SYSTEM_PAUSE | QUARANTINE | CHALLENGE_PENDING).
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
    basicAuth:
      type: http
      scheme: basic
...